Extract Msf::Module::Rank
MSP-11126
